Updates to release 2.2.7
It was updated the download links, checksum and commits for all needed
code source as well as it was added code to download the new Wings3D
dependencies (libigl and eigen).
Because the Wings build script was not able to download these dependencies,
it was needed to trick flatpak by downloading these git repositories before
they be processed by the Wings sandbox. The files were moved to a temporary
folder under $FLATPAK_DEST/_deps which was moved to Wings sandbox
later.

NOTE: Update all necessary data to build Wings3D v2.2.7